+++ eras/allegra/impl/testlib/Test/Cardano/Ledger/Allegra/Era.hs
@@ -29,8 +29,6 @@ instance EraTest AllegraEra where
 
   accountsFromAccountsMap = shelleyAccountsFromAccountsMap
 
-  accountsToUMap = shelleyAccountsToUMap
-
 instance ShelleyEraTest AllegraEra
 
 instance AllegraEraTest AllegraEra