+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Era.hs
@@ -7,21 +7,16 @@ module Test.Cardano.Ledger.Conway.Era (
   module Test.Cardano.Ledger.Babbage.Era,
   ConwayEraTest,
   mkConwayTestAccountState,
-  conwayAccountsToUMap,
 ) where
 
 import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusTxInfo)
-import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Coin
 import Cardano.Ledger.Conway
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Governance
 import Cardano.Ledger.Conway.State
 import Cardano.Ledger.Plutus (Language (..))
-import Cardano.Ledger.UMap
 import Data.Coerce
-import qualified Data.Map.Strict as Map
-import qualified Data.Set as Set
 import Lens.Micro
 import Test.Cardano.Ledger.Babbage.Era
 import Test.Cardano.Ledger.Conway.Arbitrary ()
@@ -44,8 +39,6 @@ instance EraTest ConwayEra where
   mkTestAccountState _mPtr = mkConwayTestAccountState
 
   accountsFromAccountsMap = coerce
-
-  accountsToUMap = conwayAccountsToUMap
 
 instance ShelleyEraTest ConwayEra
 
@@ -71,17 +64,3 @@ mkConwayTestAccountState deposit mStakePool mDRep =
   mkConwayAccountState deposit
     & stakePoolDelegationAccountStateL .~ mStakePool
     & dRepDelegationAccountStateL .~ mDRep
-
-conwayAccountsToUMap :: ConwayEraAccounts era => Accounts era -> UMap
-conwayAccountsToUMap accounts =
-  UMap
-    { umElems = Map.map toUMElem (accounts ^. accountsMapL)
-    , umPtrs = Map.empty
-    }
-  where
-    toUMElem accountState =
-      UMElem
-        (SJust (RDPair (accountState ^. balanceAccountStateL) (accountState ^. depositAccountStateL)))
-        Set.empty
-        (maybeToStrictMaybe (accountState ^. stakePoolDelegationAccountStateL))
-        (maybeToStrictMaybe (accountState ^. dRepDelegationAccountStateL))