+++ eras/mary/impl/testlib/Test/Cardano/Ledger/Mary/Era.hs
@@ -26,8 +26,6 @@ instance EraTest MaryEra where
 
   accountsFromAccountsMap = shelleyAccountsFromAccountsMap
 
-  accountsToUMap = shelleyAccountsToUMap
-
 instance ShelleyEraTest MaryEra
 
 instance AllegraEraTest MaryEra