+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Era.hs
@@ -1,13 +1,11 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
 
 module Test.Cardano.Ledger.Era (
   EraTest (..),
   registerTestAccount,
-  accountsFromUMap,
 ) where
 
 import Cardano.Ledger.BaseTypes
@@ -17,12 +15,9 @@ import Cardano.Ledger.Credential
 import Cardano.Ledger.Genesis
 import Cardano.Ledger.Plutus (CostModels)
 import Cardano.Ledger.State
-import Cardano.Ledger.UMap
 import Data.Aeson (FromJSON, ToJSON)
-import Data.Default (def)
 import Data.Functor.Identity
 import qualified Data.Map.Strict as Map
-import qualified Data.Set as Set
 import Data.Typeable
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.Arbitrary ()
@@ -106,8 +101,6 @@ class
 
   accountsFromAccountsMap :: Map.Map (Credential 'Staking) (AccountState era) -> Accounts era
 
-  accountsToUMap :: Accounts era -> UMap
-
 -- | This is a helper function that uses `mkTestAccountState` to register an account.
 registerTestAccount ::
   (HasCallStack, EraTest era) =>
@@ -120,22 +113,3 @@ registerTestAccount ::
   Accounts era
 registerTestAccount cred mPtr deposit mStakePool mDRep =
   addAccountState cred (mkTestAccountState mPtr deposit mStakePool mDRep)
-
--- This is a temporary converter, which is used to test some functionality until UMap is completely removed
-accountsFromUMap :: (EraTest era, HasCallStack) => UMap -> Accounts era
-accountsFromUMap umap =
-  Map.foldrWithKey' register def (umElems umap)
-  where
-    register cred (UMElem mRD ptrSet smStakePool smDRep) =
-      case mRD of
-        SNothing -> error "Invalid UMap state: missing RDPair"
-        SJust RDPair {rdReward, rdDeposit} ->
-          let mPtr =
-                case Set.toList ptrSet of
-                  [] -> Nothing
-                  [ptr] -> Just ptr
-                  ptrs -> error $ "Invalid UMap state: Can't have more than one pointer: " <> show ptrs
-              mStakePool = strictMaybeToMaybe smStakePool
-              mDRep = strictMaybeToMaybe smDRep
-           in addToBalanceAccounts (Map.singleton cred rdReward)
-                . registerTestAccount cred mPtr rdDeposit mStakePool mDRep